refactor(subm): factor out Discord Guild API
